requirements.txt
amqp==2.5.2
billiard==3.6.1.0
celery==4.4.0
certifi==2019.11.28
cffi==1.13.2
chardet==3.0.4
Click==7.0
cryptography==2.8
Deprecated==1.2.7
Django==2.2.9
django-admin-tools==0.9.0
django-cors-headers==3.2.1
django-json-field==0.5.7
django-matrix-field==0.2
djangorestframework==3.11.0
GDAL==2.4.0
geojson==2.5.0
geomet==0.2.1.post1
gitdb2==2.0.6
GitPython==3.0.5
html2text==2020.1.16
idna==2.8
importlib-metadata==1.4.0
kombu==4.6.7
more-itertools==8.1.0
numpy==1.18.1
ogr==0.9.0
opencv-python==4.1.2.30
pkg-resources==0.0.0
psycopg2==2.8.4
pycparser==2.19
PyGithub==1.45
PyJWT==1.7.1
python-dateutil==2.8.1
python-gitlab==1.15.0
pytz==2019.3
PyYAML==5.3
requests==2.22.0
sentinelsat==0.13
six==1.14.0
smmap2==2.0.5
sqlparse==0.3.0
tqdm==4.41.1
urllib3==1.25.7
vine==1.3.0
wrapt==1.11.2
zipp==2.0.0

